Create new Game. Added game ( bad logic ). Can Add new game into database. Error with Ajax-request to controller method. Without adding any match history (step and hands ).

+++ BlackJack.Entities/Card.cs
@@ -9,11 +9,11 @@ namespace BlackJack.Entities
 {
     public class Card
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
 
         public CardSuit CardSuit { get; set; }
         public CardNumber CardNumber { get; set; }
-        public int CardScore { get; set; }
+        public long CardScore { get; set; }
         public string CardName { get; set; }
 
         //Many-to-Many with PlayerHand
