+++ BlackJack.Entities/PlayerHand.cs
@@ -8,16 +8,16 @@ namespace BlackJack.Entities
 {
     public class PlayerHand
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
 
         public int Score { get; set; }
         public int Cash { get; set; }
         public int CardPoints { get; set; }
 
-        public int PlayerID { get; set; }
+        public long PlayerID { get; set; }
         public virtual User User { get; set; }
 
-        public int StepID { get; set; }
+        public long StepID { get; set; }
         public virtual Step Step { get; set; }
 
         public virtual ICollection<Card> Cards { get; set; }
